lec2 added

+++ src/main/java/org/example/sem1/HotBeverages.java
@@ -2,18 +2,18 @@ package org.example.sem1;
 
 public class HotBeverages extends Product {
 
-    private int temperature;
+    private Long temperature;
 
-    public HotBeverages(String kind, double price, Long id, int temperature) {
+    public HotBeverages(String kind, double price, Long id, Long temperature) {
         super(kind, price, id);
             this.temperature = temperature;
     }
 
-    public int getTemperature() {
+    public Long getTemperature() {
         return temperature;
     }
 
-    public void setTemperature(int temperature) {
+    public void setTemperature(Long temperature) {
         this.temperature = temperature;
     }
 }
